Remove running indentation from auth/token handlers
By using `EitherT (Handler b v ()) m a`, we can easily chain a series of
actions together, and be optimistic that they will succeed. This lets us
move the failing cases out of the main function body for clarity, and remove
the running indentation.

Also changes all backend stuff to run in IO, not MonadIO, for symmetry with
the IAuthBackend stuff in Snap.