Heavily rewriting auth grant/access token error handling
The OAuth specification is very specific that there are some cases
in the authorization grant flow that errors should be presented to
the resource owner, and sometimes should be redirected to the client.

As such, the ParamParser isn't really suitable for this flow -
because there are two phases of parsing. As such, I've ripped
authorizationRequest apart and moved most of the parsing into there.

requestToken is yet to be implemented, and currently commented out.